chore: move asset models into asset module
